+++ RELEASE_NOTES.rst
@@ -35,6 +35,8 @@ ARTIQ-4
 * ``artiq.coredevice.dds`` has been renamed to ``artiq.coredevice.ad9914`` and
   simplified. DDS batch mode is no longer supported. The ``core_dds`` device
   is no longer necessary.
+* The configuration entry ``startup_clock`` is renamed ``rtio_clock``. Switching
+  clocks dynamically (i.e. without device restart) is no longer supported.
 
 
 ARTIQ-3